Redesigned and refactored booking card
