+++ src/components/booking_card/card_element.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const Label = ({ children }: { children: any }) => (
+  <span className="block text-sm uppercase text-gray-400 font-medium leading-none pb-2 dark:text-gray-300">
+    {children}
+  </span>
+);
+
+const Title = ({ children }: { children: any }) => (
+  <p className="text-black-900 text-headline dark:text-white">{children}</p>
+);
+
+export { Label, Title };
