+++ src/components/buttons.tsx
@@ -6,7 +6,7 @@ const TaskButton = ({ onClick, actionText, icon }: any) => {
   return (
     <button
       onClick={onClick}
-      className="relative h-8 flex items-center justify-center ml-2 font-medium text-sm text-gray-400 uppercase leading-none rounded-[3rem] py-2 pl-4 pr-2 hover:bg-gray-100 transition-all ease-in-out duration-200"
+      className="relative flex items-center justify-center ml-2 text-sm uppercase text-gray-400 font-medium leading-none rounded-[3rem] py-3 pl-4 pr-3 hover:bg-gray-100 transition-all ease-in-out duration-200"
     >
       <span className="mr-2">{actionText}</span>{" "}
       <i className={iconClassName}></i>
